+++ CHANGELOG.md
@@ -17,6 +17,8 @@ Versioning](https://semver.org/spec/v2.0.0.html).
 ### Changed
 
 - Update cunnigham to 2.0.0
+- Update frontend components OrderStateMessage with new computed state 
+  'waiting signature'
 - Update SaleTunnel to add certificate product course run information.
 - Update SaleTunnel to display product instructions.
 - When a course have multiple seller organizations, CourseGlimpse now display